refactor(routing): loading.tsx 제거 및 URL 파라미터 클라이언트 내재화
- app/loading.tsx 삭제: 루트 Suspense fallback이 /me, /notices 등 전체 라우트에
  적용돼 페이지 이동 시 메인 페이지 스켈레톤이 노출되는 문제 근본 해결
- ChallengesView가 useSearchParams()로 URL 파라미터 직접 읽음: 필터 변경마다
  불필요한 서버 왕복 제거
- app/page.tsx: async 제거, searchParams 의존 제거
- lib/queries/params.ts 신설: ListedProblem·파서 함수를 DB 의존 없는 파일로 분리,
  ChallengesView 클라이언트 번들에서 postgres 체인 차단
